fix: complete TextTransformation migration, remove NaturalLanguage project (#352)
- Remove DocGeneration.Core.NaturalLanguage from solution and delete project
- Remove DocGeneration.Core.NaturalLanguage.Tests and delete project
- Remove NaturalLanguage project references from Annotations, HorizontalArticles, RawTools
- Delete orphaned TextCleanupFixture (no IClassFixture consumers remain)
- Port 15 EnsureEndsPeriod tests to TextNormalizerTests (previously only in NaturalLanguage.Tests)
- Update comments referencing TextCleanup to describe current architecture
- All production code already uses TransformationEngine/TextNormalizer
- Zero behavioral regression

Closes #352
