+++ mcp-tools/DocGeneration.Steps.HorizontalArticles/DocGeneration.Steps.HorizontalArticles.csproj
@@ -15,7 +15,6 @@
     <ProjectReference Include="..\DocGeneration.Steps.AnnotationsParametersRaw.Annotations\DocGeneration.Steps.AnnotationsParametersRaw.Annotations.csproj" />
     <ProjectReference Include="..\..\shared\DocGeneration.Core.Shared\DocGeneration.Core.Shared.csproj" />
     <ProjectReference Include="..\..\shared\DocGeneration.Core.GenerativeAI\DocGeneration.Core.GenerativeAI.csproj" />
-    <ProjectReference Include="..\..\shared\DocGeneration.Core.NaturalLanguage\DocGeneration.Core.NaturalLanguage.csproj" />
     <ProjectReference Include="..\..\shared\DocGeneration.Core.TextTransformation\DocGeneration.Core.TextTransformation.csproj" />
     <ProjectReference Include="..\..\shared\DocGeneration.Core.TemplateEngine\DocGeneration.Core.TemplateEngine.csproj" />
   </ItemGroup>
