--- shared/DocGeneration.Core.NaturalLanguage.Tests/TestData/test-nl-parameters.json
@@ -1,14 +1,0 @@
-[
-  {
-    "Parameter": "auth-method",
-    "NaturalLanguage": "Authentication method"
-  },
-  {
-    "Parameter": "cluster-uri",
-    "NaturalLanguage": "Cluster URI"
-  },
-  {
-    "Parameter": "nodepool",
-    "NaturalLanguage": "Node pool"
-  }
-]
