--- shared/DocGeneration.Core.NaturalLanguage.Tests/TestData/test-static-text-replacement.json
@@ -1,58 +1,0 @@
-[
-  {
-    "Parameter": "Azure Active Directory",
-    "NaturalLanguage": "Microsoft Entra ID"
-  },
-  {
-    "Parameter": "Azure AD",
-    "NaturalLanguage": "Microsoft Entra ID"
-  },
-  {
-    "Parameter": "CosmosDB",
-    "NaturalLanguage": "Azure Cosmos DB"
-  },
-  {
-    "Parameter": "VMSS",
-    "NaturalLanguage": "Virtual machine scale set (VMSS)"
-  },
-  {
-    "Parameter": "whitelist",
-    "NaturalLanguage": "allowlist"
-  },
-  {
-    "Parameter": "blacklist",
-    "NaturalLanguage": "blocklist"
-  },
-  {
-    "Parameter": "leverage",
-    "NaturalLanguage": "use"
-  },
-  {
-    "Parameter": "utilize",
-    "NaturalLanguage": "use"
-  },
-  {
-    "Parameter": "e.g.",
-    "NaturalLanguage": "for example"
-  },
-  {
-    "Parameter": "etc.",
-    "NaturalLanguage": "and more"
-  },
-  {
-    "Parameter": "master branch",
-    "NaturalLanguage": "main branch"
-  },
-  {
-    "Parameter": "sanity check",
-    "NaturalLanguage": "validation check"
-  },
-  {
-    "Parameter": "prior to",
-    "NaturalLanguage": "before"
-  },
-  {
-    "Parameter": "in order to",
-    "NaturalLanguage": "to"
-  }
-]
